fix: removed services, added use cases

+++ src/main/kotlin/concentric_circles/clean_architecture/useCase/RemoveProductFromInventoryUseCase.kt
@@ -2,6 +2,7 @@ package concentric_circles.clean_architecture.useCase
 
 import concentric_circles.clean_architecture.repository.InventoryRepository
 import org.springframework.stereotype.Service
+import java.util.UUID
 
 @Service
 class RemoveProductFromInventoryUseCase (
@@ -9,12 +10,17 @@ class RemoveProductFromInventoryUseCase (
     var productExistsUseCase: CheckIfProductExistsUseCase
 ) {
 
-    fun removeProductFromInventory (productName: String) : Boolean {
+    fun removeProductFromInventory (productId: UUID) : Boolean {
 
-        val existingProduct = productExistsUseCase.checkIfProductExists(productName)
-            ?: throw Exception("Product '${productName}' does not exist.")
+        val existingProduct = productExistsUseCase.checkIfProductExists(productId)
+//            ?: throw Exception("Product '${existingProduct.name}' does not exist.")
 
-        return inventoryRepository.deleteInventoryByProductId(existingProduct.productId)
+        if (existingProduct != null) {
+            inventoryRepository.deleteInventoryByProductId(existingProduct.productId)
+            return true
+        }
+
+        return false
     }
 
 }
